Made remotedb_webhook pass PHPStan.

+++ modules/remotedb_webhook/src/Form/WebhookDisable.php
@@ -3,6 +3,7 @@
 namespace Drupal\remotedb_webhook\Form;
 
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\remotedb\Entity\RemotedbInterface;
 
 /**
  * Removes a webhook url.
@@ -29,6 +30,9 @@ class WebhookDisable extends OperationActionBase {
    * {@inheritdoc}
    */
   public function submitForm(array &$form, FormStateInterface $form_state): void {
+    if (!$this->entity instanceof RemotedbInterface) {
+      throw new \LogicException(sprintf('Entity is not of the correct type. Should be a %s but it is a %s.', RemotedbInterface::class, get_class($this->entity)));
+    }
     $this->webhookService->delete($this->entity);
 
     $success = !$this->webhookService->exists($this->entity);
